… src/benchmarks/micro/libraries/Common/Chinese.txt
… src/benchmarks/micro/libraries/Common/Cyrillic.txt
… src/benchmarks/micro/libraries/Common/EnglishAllAscii.txt
… src/benchmarks/micro/libraries/Common/EnglishMostlyAscii.txt
… src/benchmarks/micro/libraries/Common/Greek.txt