The purpose of this PR is to verify the performance of the recent AdvSimd.Arm64 improvements [added to Utf16Utility.Validation.cs](https://github.com/dotnet/runtime/pull/39050).

The `UTF8Encoding.GetByteCount` method reaches the Utf16Utility.Validation.cs lines [82](https://github.com/dotnet/runtime/blob/562763b5498bf53b68517b68221a89ad80ef6d16/src/libraries/System.Private.CoreLib/src/System/Text/Unicode/Utf16Utility.Validation.cs#L82), [150](https://github.com/dotnet/runtime/blob/562763b5498bf53b68517b68221a89ad80ef6d16/src/libraries/System.Private.CoreLib/src/System/Text/Unicode/Utf16Utility.Validation.cs#L150) and [190](https://github.com/dotnet/runtime/blob/562763b5498bf53b68517b68221a89ad80ef6d16/src/libraries/System.Private.CoreLib/src/System/Text/Unicode/Utf16Utility.Validation.cs#L190), when called with the contents of the [recently added text files](https://github.com/dotnet/performance/commit/ba1df98a6b22d04371239e6ce2620f17b58c6e39) containing long unicode strings.

cc @kunalspathak @pgovind @adamsitnik