@carlossanlop 2020-08-19
Results:

### SSE2 OFF vs SSE2 ON

Off: `$Env:COMPlus_EnableSSE2=0`
On: `$Env:COMPlus_EnableSSE2=1`

```
❯ D:\runtime\.dotnet\dotnet.exe run --base C:\users\carlos\Desktop\sse_off\ --diff C:\users\carlos\Desktop\sse_on\ --threshold 0.001%
summary:
better: 20, geomean: 1.854
total diff: 20

No Slower results for the provided threshold = 0.001% and noise filter = 0.3ns.

| Faster                                                                | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Text.Perf_Utf8Encoding.GetByteCount(Input: Greek)              |     10.49 |        154142.40 |         14700.10 |         |
| System.Text.Perf_Utf8Encoding.GetByteCount(Input: Cyrillic)           |      8.97 |        104052.54 |         11601.99 |         |
| System.Text.Perf_Utf8Encoding.GetByteCount(Input: Chinese)            |      5.33 |         72346.37 |         13583.33 |         |
| System.Text.Perf_Utf8Encoding.GetByteCount(Input: EnglishAllAscii)    |      4.04 |         29311.31 |          7260.16 |         |
| System.Text.Perf_Utf8Encoding.GetByteCount(Input: EnglishMostlyAscii) |      3.35 |         98971.88 |         29549.56 |         |
```


### AdvSimd OFF vs AdvSimd ON

Off: `export COMPlus_EnableArm64AdvSimd=0`
On: `export COMPlus_EnableArm64AdvSimd=1`

```
root@calopearm:/home/calope/performance/src/tools/ResultsComparer# dn run -c release --base /home/calope/arm_off/ --diff /home/calope/arm_on/ --threshold 0.001%summary:
better: 13, geomean: 1.772
total diff: 13

No Slower results for the provided threshold = 0.001% and noise filter = 0.3ns.

| Faster                                                                | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Text.Perf_Utf8Encoding.GetByteCount(Input: Greek)              |      3.50 |        295751.19 |         84604.67 |         |
| System.Text.Perf_Utf8Encoding.GetByteCount(Input: Cyrillic)           |      2.76 |        225379.32 |         81806.99 | several?|
| System.Text.Perf_Utf8Encoding.GetByteCount(Input: Chinese)            |      2.52 |        209528.09 |         83099.90 |         |
| System.Text.Perf_Utf8Encoding.GetByteCount(Input: EnglishAllAscii)    |      1.56 |         64962.94 |         41614.67 |         |
| System.Text.Perf_Utf8Encoding.GetByteCount(Input: EnglishMostlyAscii) |      1.16 |        228289.04 |        196492.73 |         |
```