@adamsitnik 2020-08-24
> I'm just wondering if this will break the mono CI too.

This one not because it's not using the forbidden constructor :D